backend/requirements.txt
# Sistema Gestionale Aziendale - Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
asyncpg==0.29.0
databases==0.8.0
supabase==2.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.3
email-validator==2.1.0
httpx==0.25.2
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3
